Move defaults into separate module
This functionality can be cleanly encapsulated into its own module.
This makes `fft_array.py` shorter and more focused.
