+++ fftarray/fft_dimension.py
@@ -9,8 +9,9 @@ from .backends.numpy import NumpyBackend
 
 from ._utils._formatting import fft_dim_table, format_n
 from ._utils._indexing import check_substepping, remap_index_check_int
+from ._utils._defaults import get_default_backend, get_default_eager
 
-from .fft_array import Space, FFTArray, get_default_backend, get_default_eager
+from .fft_array import Space, FFTArray
 
 @dataclass
 class FFTDimension:
